+++ sdk/cosmosdb/cosmos/src/client/Container/Container.ts
@@ -456,7 +456,12 @@ export class Container {
       }, this.clientContext);
     }
   }
-
+  /**
+   * This function handles the scenario where a container is deleted(say from different Client) and recreated with same Id but with different client encryption policy.
+   * The idea is to have the container Rid cached and sent out as part of RequestOptions with Container Rid set in "x-ms-cosmos-intended-collection-rid" header.
+   * So, when the container being referenced here gets recreated we would end up with a stale encryption settings and container Rid and this would result in BadRequest (and a substatus 1024).
+   * This would allow us to refresh the encryption settings and Container Rid, on the premise that the container recreated could possibly be configured with a new encryption policy.
+   */
   async ThrowIfRequestNeedsARetryPostPolicyRefresh(errorResponse: any): Promise<void> {
     const key = this.database.id + "/" + this.id;
     const encryptionSetting =
