+++ sdk/cosmosdb/cosmos/tsconfig.strict.json
@@ -154,6 +154,7 @@
     "src/utils/supportedQueryFeaturesBuilder.ts",
     "src/encryption/*",
     "test/public/common/TestHelpers.ts",
+    "test/public/common/encryptionTestHelpers.ts",
     "test/internal/session.spec.ts",
     "test/internal/unit/auth.spec.ts",
     "test/internal/unit/defaultQueryExecutionContext.spec.ts",
@@ -176,6 +177,8 @@
     "test/internal/unit/timeoutFailoverRetryPolicy.spec.ts",
     "test/internal/unit/nonStreamingOrderByMap.spec.ts",
     "test/internal/unit/utils/supportedQueryFeaturesBuilder.spec.ts",
+    "test/internal/unit/utils/strings.spec.ts",
+    "test/internal/unit/encryption/*",
     "test/public/common/BaselineTest.PathParser.ts",
     "test/public/common/TestData.ts",
     "test/public/common/setup.ts",
@@ -205,7 +208,6 @@
     "test/public/functional/user.spec.ts",
     "test/public/functional/queryIterator.spec.ts",
     "test/public/functional/clientSideEncryption.spec.ts",
-    "test/public/functional/dotnetcse.spec.ts",
     "test/public/integration/aggregateQuery.spec.ts",
     "test/public/integration/authorization.spec.ts",
     "test/public/integration/changeFeed.spec.ts",
